Put MyStack into its own file, removed from StackTrio file
