+++ Python/chapter03/p01_three_in_one/miguelHx.py
@@ -18,109 +18,6 @@ class StackNode(Generic[T]):
     next: 'Optional[StackNode[T]]'
 
 
-class MyStack:
-    """Stack data structure implementation.
-
-    Uses LIFO (last-in first-out) ordering.
-    The most recent item added to the stack is
-    the first removed.  Traversal is top to bottom.
-    """
-
-    def __init__(self):
-        self.top: Optional[StackNode[T]] = None # top is a pointer to StackNode object
-        self.size: int = 0
-        self.index: int = -1 # tracking current node for iterator use
-        self.current_node: Optional[StackNode[T]] = self.top
-        return
-    
-    def pop(self) -> T:
-        """
-        Removes the top item from the stack
-
-        Raises:
-            IndexError: raised when pop is attempted on empty stack
-
-        Returns:
-            int: The data at the top of the stack
-        """
-        if self.top is None:
-            raise IndexError('Stack is Empty.')
-        item = self.top.data
-        self.top = self.top.next
-        self.current_node = self.top
-        self.size -= 1
-        return item
-    
-    def push(self, item: int) -> None:
-        """
-        Adds an item to the top of the stack
-        Args:
-            item (int): data we want at the top of stack
-        """
-        t = StackNode(item, None)
-        t.next = self.top
-        self.top = t
-        self.current_node = self.top
-        self.size += 1
-
-    def peek(self) -> int:
-        """
-        Returns data at the top of the stack
-
-        Raises:
-            IndexError: [description]
-
-        Returns:
-            int: the value at the top of the stack
-        """
-        if self.top is None:
-            raise IndexError('Stack is Empty')
-        return self.top.data
-    
-    def __iter__(self) -> Iterator:
-        """
-        Builds a list of the current stack state.
-        For example, given the following stack:
-            3 -> 2 -> 1, where 3 is the top,
-        Expect:
-            [3, 2, 1]
-        Returns:
-            List[int]: list of integers
-        """
-        return self
-    
-    def __next__(self) -> T:
-        self.index += 1
-        if self.index == self.size or self.current_node is None:
-            self.index = -1
-            self.current_node = self.top
-            raise StopIteration
-        n: T = self.current_node.data
-        self.current_node = self.current_node.next
-        return n
-    
-    def __bool__(self) -> bool:
-        """
-        Returns True if and only if the
-        stack is empty, otherwise False.
-
-        Returns:
-            bool: True if empty, False otherwise
-        """
-        return self.size == 0
-    
-    def __str__(self):
-        if self.size == 0:
-            return '<Empty>'
-        values = []
-        n = self.top
-        while n.next is not None:
-            values.append(str(n.data))
-            n = n.next
-        values.append(str(n.data))
-        return '->'.join(values)
-
-
 class StackTrio:
 
     def __init__(self, stack_capacity = 7):
@@ -229,7 +126,6 @@ class StackTrio:
             self.stack_info[stack_id]['size'] -= 1
             return val_before_pop
 
-        
         # otherwise, stack size is greater than 2, and we can decrement
         # top index and next index by 1
         self.stack_info[stack_id]['top_index'] -= 1
@@ -256,65 +152,6 @@ class StackTrio:
         d = copy.deepcopy(self.stack_info)
         d['values'] = str(self.values)
         return str(d)
-        
-
-
-        
-
-
-class TestMyStack(unittest.TestCase):
-    def test_stack_push(self):
-        s = MyStack()
-        self.assertEqual(s.size, 0)
-        self.assertEqual(s.top, None)
-        s.push(2)
-        self.assertEqual(s.size, 1)
-        self.assertEqual(s.top.data, 2)
-        self.assertEqual(s.top.next, None)
-        s.push(3)
-        self.assertEqual(s.size, 2)
-        self.assertEqual(s.top.data, 3)
-        self.assertEqual(s.top.next.data, 2)
-        s.push(4)
-        self.assertEqual(s.size, 3)
-        self.assertEqual(s.top.data, 4)
-        self.assertEqual(s.top.next.data, 3)
-
-        l = list(s)
-        self.assertEqual(l, [4, 3, 2])
-    
-    def test_stack_peek(self):
-        s = MyStack()
-        with self.assertRaises(IndexError):
-            s.peek()
-        s.push(1)
-        s.push(2)
-        s.push(99)
-        top_val = s.peek()
-        self.assertEqual(top_val, 99)
-    
-    def test_stack_is_empty(self):
-        s = MyStack()
-        self.assertTrue(s)
-        s.push(7)
-        self.assertFalse(s)
-    
-    def test_stack_pop(self):
-        # first case, attempt to pop an empty stack
-        s = MyStack()
-        with self.assertRaises(IndexError):
-            s.pop()
-        s.push(1)
-        s.push(2)
-        s.push(3)
-        # size is 3
-        self.assertEqual(list(s), [3, 2, 1])
-        val = s.pop()
-        self.assertEqual(val, 3)
-        self.assertEqual(s.size, 2) # size should now be 2
-        self.assertEqual(list(s), [2, 1])
-
-
 
 
 class TestThreeInOne(unittest.TestCase):
